Allow user to remap modifiers
This adds support for mapping physical modifier keys, as reported by
the Kitty Keyboard Protocol, to different Emacs modifiers.  It is
modeled after the behavior of the `mac-command-modifer',
'mac-option-modifier' and related variables.

By default all the new `kpp-*-modifier' variables map 1-to-1 with the
named modifier.  This introduces a change from prior `kpp.el' behavior
where <Alt> and <Meta> both mapped to <Meta>.  Currently the
`kpp-alt-modifer' maps to <Alt>.  This may be changed by a simple:

  (setq kpp-alt-modifer 'meta)

The motivation for this change was that in Emacs on macOS my
Command (⌘) key is set to produce <Meta> and I wanted the same
behavior with KPP in a terminal.