Don't concatenate the ?u and ?~ lists when retrieving the keycode representation
The two lists are not disjunct. For example '13' maps to "<return>" in
the ?u list, and "<f3>" in the ?~ list.

This fixes an actual issue, where F3 is incorrectly mapped to <return>
in terminal emulators implementing the latest version of the kkp
specification, where the legacy sequence for F3 is no longer
allowed (due to it conflicting with cursor position reports).